finishing uo changes before i finally make a release

+++ src/clientsideApp/src/main/java/chessengine/PgnFunctions.java
@@ -1,30 +1,20 @@
 package chessengine;
 
-import java.util.*;
+import java.util.Arrays;
+import java.util.LinkedList;
+import java.util.List;
 
 public class PgnFunctions {
-
-
 
 
     // all related to creating a position out of a pgn
 
 
-
-
-
-
-
-
-
-
-
-
-    private static List<String> parseMoveInfo(String pgnInfo){
+    private static List<String> parseMoveInfo(String pgnInfo) {
         return new LinkedList<String>(Arrays.asList(pgnInfo.split("\\[.*?\\]")));
     }
 
-    public static String[] splitPgn(String pgn){
+    public static String[] splitPgn(String pgn) {
 //        String infoPart = "";
 //        String moveTextPart = "";
 //        Pattern pattern = Pattern.compile("^(\\[.*?\\])\\s*((?:\\d+\\.\\s.*)+)$", Pattern.MULTILINE);
@@ -43,36 +33,35 @@ public class PgnFunctions {
 //        }
 //        return new String[]{infoPart,moveTextPart};
         int moveTextIndex = pgn.indexOf("1.");
-        return new String[]{pgn.substring(0,moveTextIndex),pgn.substring(moveTextIndex)};
+        return new String[]{pgn.substring(0, moveTextIndex), pgn.substring(moveTextIndex)};
     }
+
     // pgn examples: 1. e4 e5 2. Nf3 Nc6 3. Bb5...
-    public static String moveToPgn(ChessMove move,ChessPosition pos,ChessStates gameStates){
+    public static String moveToPgn(ChessMove move, ChessPosition pos, ChessStates gameStates) {
         StringBuilder sb1 = new StringBuilder();
-        if(move.isCastleMove()){
-            int xDiff = move.getNewX()-move.getOldX();
-            if(xDiff > 0){
+        if (move.isCastleMove()) {
+            int xDiff = move.getNewX() - move.getOldX();
+            if (xDiff > 0) {
                 // short castle
                 return "O-O";
-            }
-            else{
+            } else {
                 // long castle
                 return "O-O-O";
             }
         }
-        String start = turnPieceIndexToStr(move.getBoardIndex(),true);
-        String ambiguityChar = AdvancedChessFunctions.getAmbigiousStr(move.getOldX(),move.getOldY(),move.getNewX(),move.getNewY(),move.getBoardIndex(),move.isWhite(),move.isEating(),pos);
+        String start = turnPieceIndexToStr(move.getBoardIndex(), true);
+        String ambiguityChar = AdvancedChessFunctions.getAmbigiousStr(move.getOldX(), move.getOldY(), move.getNewX(), move.getNewY(), move.getBoardIndex(), move.isWhite(), move.isEating(), pos);
         String eatingChar = move.isEating() || move.isEnPassant() ? "x" : "";
         char xChar = turnIntToFileStr(move.getNewX());
         // flip y
-        char yChar = intToChar(7-move.getNewY()+1);
+        char yChar = intToChar(7 - move.getNewY() + 1);
         String promoChar = move.isPawnPromo() ? "=" : "";
-        String promoTypeChar = move.isPawnPromo() ? turnPieceIndexToStr(move.getPromoIndx(),true) : "";
-        String checkedChar = AdvancedChessFunctions.isChecked(!move.isWhite(),pos.board) ? "+" : "";
-        if(AdvancedChessFunctions.isAnyNotMovePossible(!move.isWhite(),pos,gameStates)){
-            if(AdvancedChessFunctions.isChecked(!move.isWhite(),pos.board)){
+        String promoTypeChar = move.isPawnPromo() ? turnPieceIndexToStr(move.getPromoIndx(), true) : "";
+        String checkedChar = AdvancedChessFunctions.isChecked(!move.isWhite(), pos.board) ? "+" : "";
+        if (AdvancedChessFunctions.isAnyNotMovePossible(!move.isWhite(), pos, gameStates)) {
+            if (AdvancedChessFunctions.isChecked(!move.isWhite(), pos.board)) {
                 checkedChar = "#";
-            }
-            else{
+            } else {
                 checkedChar = "";
             }
 
@@ -81,19 +70,18 @@ public class PgnFunctions {
 
     }
 
-    public static String moveToPgn(ChessPosition p,ChessStates gameStates){
-        return moveToPgn(p.getMoveThatCreatedThis(),p,gameStates);
+    public static String moveToPgn(ChessPosition p, ChessStates gameStates) {
+        return moveToPgn(p.getMoveThatCreatedThis(), p, gameStates);
 
     }
 
-    public static String invertPgn(String pgn){
+    public static String invertPgn(String pgn) {
         StringBuilder sb = new StringBuilder();
-        for(int i = 0;i<pgn.length();i++){
+        for (int i = 0; i < pgn.length(); i++) {
             char c = pgn.charAt(i);
-            if(Character.isDigit(c)){
-                sb.append(intToChar(9-Integer.parseInt(String.valueOf(c))));
-            }
-            else {
+            if (Character.isDigit(c)) {
+                sb.append(intToChar(9 - Integer.parseInt(String.valueOf(c))));
+            } else {
                 sb.append(c);
             }
         }
@@ -101,28 +89,23 @@ public class PgnFunctions {
     }
 
 
-
-    public static char intToChar(int i){
-        return (char) (i+48);
+    public static char intToChar(int i) {
+        return (char) (i + 48);
     }
 
-    public static int turnFileStrToInt(char c){
-        int ascii = (int) c;
-        return ascii-97;
+    public static int turnFileStrToInt(char c) {
+        int ascii = c;
+        return ascii - 97;
     }
 
-    public static char turnIntToFileStr(int i){
-        char ascii = (char)i;
-        return (char) (ascii+97);
+    public static char turnIntToFileStr(int i) {
+        char ascii = (char) i;
+        return (char) (ascii + 97);
     }
 
 
-
-
-
-
-    public static int turnPgnPieceToPieceIndex(char c){
-        switch (c){
+    public static int turnPgnPieceToPieceIndex(char c) {
+        switch (c) {
             case 'K':
                 return 5;
             case 'Q':
@@ -140,8 +123,8 @@ public class PgnFunctions {
 
     }
 
-    public static String turnPieceIndexToStr(int c,boolean isPgn){
-        switch (c){
+    public static String turnPieceIndexToStr(int c, boolean isPgn) {
+        switch (c) {
             case ChessConstants.KINGINDEX:
                 return "K";
             case ChessConstants.QUEENINDEX:
@@ -154,70 +137,67 @@ public class PgnFunctions {
                 return "N";
         }
         // if none match then it is a pawn(empty char)
-        if(isPgn){
+        if (isPgn) {
             return "";
-        }
-        else{
+        } else {
             return "P";
         }
 
 
     }
-//
-    public static String positionToFEN(ChessPosition pos, ChessStates gameStateForPos,boolean isWhiteTurn){
+
+    //
+    public static String positionToFEN(ChessPosition pos, ChessStates gameStateForPos, boolean isWhiteTurn) {
         // example FEN: 8/5k2/3p4/1p1Pp2p/pP2Pp1P/P4P1K/8/8 b - - 99 50
         // go through each file in each row
         int numHalfMovesSinceCheckOrPawnMove = gameStateForPos.getMovesSinceNoCheckOrNoPawn();
-        int numFullMoves = gameStateForPos.getCurrentIndex()/2;
+        int numFullMoves = gameStateForPos.getCurrentIndex() / 2;
         StringBuilder fenBuilder = new StringBuilder();
-        for(int row  = 0;row<8;row++){
+        for (int row = 0; row < 8; row++) {
             int emptySquareCount = 0;
-            for(int file = 0;file<8;file++){
-                boolean[] boardInfo = GeneralChessFunctions.checkIfContains(file,row,pos.board,"tofen");
-                if(boardInfo[0]){
+            for (int file = 0; file < 8; file++) {
+                boolean[] boardInfo = GeneralChessFunctions.checkIfContains(file, row, pos.board, "tofen");
+                if (boardInfo[0]) {
                     // means there is a piece there
                     // first put the number of empty squares before this
-                    if(emptySquareCount > 0){
+                    if (emptySquareCount > 0) {
                         fenBuilder.append(emptySquareCount);
                     }
                     emptySquareCount = 0;
                     // now we put the string representing the piece
-                    int pieceIndex = GeneralChessFunctions.getBoardWithPiece(file,row,boardInfo[1],pos.board);
+                    int pieceIndex = GeneralChessFunctions.getBoardWithPiece(file, row, boardInfo[1], pos.board);
                     String pieceString;
-                    if(boardInfo[1]){
+                    if (boardInfo[1]) {
                         // means a white piece
-                        pieceString = turnPieceIndexToStr(pieceIndex,false);
+                        pieceString = turnPieceIndexToStr(pieceIndex, false);
 
-                    }
-                    else{
-                        pieceString = turnPieceIndexToStr(pieceIndex,false).toLowerCase();
+                    } else {
+                        pieceString = turnPieceIndexToStr(pieceIndex, false).toLowerCase();
 
                     }
                     fenBuilder.append(pieceString);
 
-                }
-                else{
+                } else {
                     // no piece so we increment empty square count;
 
                     emptySquareCount++;
                 }
             }
             // flush any empty squares
-            if(emptySquareCount > 0){
+            if (emptySquareCount > 0) {
                 fenBuilder.append(emptySquareCount);
             }
 
 
             // add a slash after every row
-            if(row != 7){
+            if (row != 7) {
                 fenBuilder.append("/");
             }
         }
         // finally add all of the extra information, such as iswhiteturn, castling rights, en passant possibilities etc
-        if(isWhiteTurn){
+        if (isWhiteTurn) {
             fenBuilder.append(" w");
-        }
-        else{
+        } else {
             fenBuilder.append(" b");
         }
         fenBuilder.append(" ");
@@ -228,10 +208,9 @@ public class PgnFunctions {
         String wqsC = gameStateForPos.isWhiteLongRookRight() && whiteCastle ? "Q" : "";
         String bksC = gameStateForPos.isWhiteShortRookRight() && blackCastle ? "k" : "";
         String bqsC = gameStateForPos.isWhiteShortRookRight() && blackCastle ? "q" : "";
-        if(wksC.isEmpty() && wqsC.isEmpty() && bksC.isEmpty() && bqsC.isEmpty()){
+        if (wksC.isEmpty() && wqsC.isEmpty() && bksC.isEmpty() && bqsC.isEmpty()) {
             fenBuilder.append("-");
-        }
-        else{
+        } else {
             fenBuilder.append(wksC);
             fenBuilder.append(wqsC);
             fenBuilder.append(bksC);
@@ -240,16 +219,16 @@ public class PgnFunctions {
         fenBuilder.append(" ");
         boolean isPassant = false;
         // now check enpassant
-        if(!pos.equals(ChessConstants.startBoardState)){
+        if (!pos.equals(ChessConstants.startBoardState)) {
             // means we arent at the very beginning as there is not an actual move for the start position
             ChessMove moveThatCreated = pos.getMoveThatCreatedThis();
-            if(moveThatCreated.getBoardIndex() == ChessConstants.PAWNINDEX){
+            if (moveThatCreated.getBoardIndex() == ChessConstants.PAWNINDEX) {
                 // pawn move so possibilty of enpassant
-                if(Math.abs(moveThatCreated.getOldY()-moveThatCreated.getNewY()) > 1){
+                if (Math.abs(moveThatCreated.getOldY() - moveThatCreated.getNewY()) > 1) {
                     // jumped 2 so means that there is a possibilty of en passant, so add to the FEN
                     isPassant = true;
                     char fileChar = turnIntToFileStr(moveThatCreated.getNewX());
-                    int midY = (moveThatCreated.getOldY() + moveThatCreated.getNewY())/2;
+                    int midY = (moveThatCreated.getOldY() + moveThatCreated.getNewY()) / 2;
 
                     fenBuilder.append(fileChar);
                     fenBuilder.append(midY);
@@ -257,7 +236,7 @@ public class PgnFunctions {
                 }
             }
         }
-        if(!isPassant){
+        if (!isPassant) {
             fenBuilder.append("-");
         }
         fenBuilder.append(" ");
@@ -266,25 +245,24 @@ public class PgnFunctions {
         fenBuilder.append(numFullMoves);
 
 
-
         return fenBuilder.toString();
     }
 
-    public static ChessMove uciToChessMove(String uci,boolean isWhite, BitBoardWrapper board){
+    public static ChessMove uciToChessMove(String uci, boolean isWhite, BitBoardWrapper board) {
         char[] c = uci.toCharArray();
         int startX = turnFileStrToInt(c[0]);
         int endX = turnFileStrToInt(c[2]);
-        int startY = 7-(Integer.parseInt(Character.toString(c[1]))-1);
-        int endY = 7-(Integer.parseInt(Character.toString(c[3]))-1);
-        int boardIndex = GeneralChessFunctions.getBoardWithPiece(startX,startY,board);
+        int startY = 7 - (Integer.parseInt(Character.toString(c[1])) - 1);
+        int endY = 7 - (Integer.parseInt(Character.toString(c[3])) - 1);
+        int boardIndex = GeneralChessFunctions.getBoardWithPiece(startX, startY, board);
         // eating (technically should check to see if its a correct piece color or throw error but nahh its fine)
-        boolean isEating = GeneralChessFunctions.checkIfContains(endX,endY,board,"ya")[0];
-        int eatingIndex = GeneralChessFunctions.getBoardWithPiece(endX,endY,board);
-        boolean isCastleMove = uci.equals("e1g1") /* white short castle */||  uci.equals("e8g8") /*black short castle*/|| uci.equals("e1c1") /* white long*/||uci.equals("e8c8")/*black short*/;
+        boolean isEating = GeneralChessFunctions.checkIfContains(endX, endY, board, "ya")[0];
+        int eatingIndex = GeneralChessFunctions.getBoardWithPiece(endX, endY, board);
+        boolean isCastleMove = uci.equals("e1g1") /* white short castle */ || uci.equals("e8g8") /*black short castle*/ || uci.equals("e1c1") /* white long*/ || uci.equals("e8c8")/*black short*/;
         int promoIndex = c.length > 4 ? turnPgnPieceToPieceIndex(Character.toUpperCase(c[4])) : ChessConstants.EMPTYINDEX;
         int backDir = isWhite ? 1 : -1;
-        boolean isEnPassant = startX != endX && boardIndex == ChessConstants.PAWNINDEX && GeneralChessFunctions.getBoardWithPiece(endX,endY+backDir,!isWhite,board) == ChessConstants.PAWNINDEX && !isEating;
-        return new ChessMove(startX,startY,endX,endY,promoIndex,boardIndex,isWhite,isCastleMove,isEating,eatingIndex,isEnPassant,false);
+        boolean isEnPassant = startX != endX && boardIndex == ChessConstants.PAWNINDEX && GeneralChessFunctions.getBoardWithPiece(endX, endY + backDir, !isWhite, board) == ChessConstants.PAWNINDEX && !isEating;
+        return new ChessMove(startX, startY, endX, endY, promoIndex, boardIndex, isWhite, isCastleMove, isEating, eatingIndex, isEnPassant, false);
 
     }
 }
